.h file change

+++ 3.Implimentation/inc/phonebook.h
@@ -1,7 +1,7 @@
 /**
- * @file header.h
- * @author your name (you@domain.com)
- * @brief 
+ * @file phonebook.h
+ * @author Gaurav_Jain (you@domain.com)
+ * @brief header file
  * @version 0.1
  * @date 2021-07-07
  * 
@@ -25,7 +25,7 @@ struct person
     char address[50];
     char father_name[35];
     char mother_name[30];
-    long int mble_no;
+    long double mble_no;
     char sex[8];
     char mail[100];
     char citision_no[20];
